Cat Beckoning
Cats can now be attracted by pinching one's fingers together within the cat's view. Also implemented the basics of a cat personality system with a host of settings that affect when or if a cat will approach these hand gestures, simulating an approachable or standoffish cat. The option to have a cat stop when touched can now be toggled.